Vietnamese translation
I'm using Pomodoro Tracker for the first time today and think it very helpful. So I'm trying to do some help as I translated it to Vietnamese which is my mother language. There are still some parts that I didn't get the chance to see it on the web, so I don't know how to translate it I hope it helps.